+++ src/main/java/com/atlassian/braid/document/MappingContext.java
@@ -1,135 +1,240 @@
 package com.atlassian.braid.document;
 
+import com.atlassian.braid.java.util.BraidObjects;
 import graphql.language.Field;
+import graphql.language.FragmentDefinition;
+import graphql.language.FragmentSpread;
+import graphql.language.InlineFragment;
 import graphql.language.ObjectTypeDefinition;
+import graphql.language.OperationDefinition;
+import graphql.language.OperationTypeDefinition;
 import graphql.schema.idl.TypeDefinitionRegistry;
 import graphql.schema.idl.TypeInfo;
 
-import java.io.Serializable;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Supplier;
+import java.util.function.Function;
+import java.util.function.Predicate;
 import java.util.stream.Stream;
 
+import static com.atlassian.braid.TypeUtils.findOperationDefinitions;
 import static com.atlassian.braid.document.Fields.getFieldAliasOrName;
 import static com.atlassian.braid.document.Fields.maybeFindObjectTypeDefinition;
 import static com.atlassian.braid.document.Fields.maybeGetTypeInfo;
+import static com.atlassian.braid.document.TypeMappers.maybeFindTypeMapper;
+import static com.atlassian.braid.java.util.BraidLists.concat;
 import static java.util.Collections.emptyList;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toList;
+import static java.util.stream.Stream.concat;
 
-final class MappingContext {
+abstract class MappingContext {
 
     private final TypeDefinitionRegistry schema;
     private final List<TypeMapper> typeMappers;
-    private final Field field;
+    private final List<FragmentDefinition> fragmentDefinitions;
 
-    private final Supplier<Optional<TypeInfo>> typeInfo;
-    private final Supplier<Optional<ObjectTypeDefinition>> objectTypeDefinition;
-    private final Supplier<List<String>> path;
+    MappingContext(MappingContext mappingContext) {
+        this(mappingContext.schema, mappingContext.typeMappers, mappingContext.fragmentDefinitions);
+    }
 
-    private MappingContext(TypeDefinitionRegistry schema,
-                           List<TypeMapper> typeMappers,
-                           List<String> parentPath,
-                           ObjectTypeDefinition parentObjectTypeDefinition,
-                           Field field) {
+    MappingContext(TypeDefinitionRegistry schema,
+                   List<TypeMapper> typeMappers,
+                   List<FragmentDefinition> fragmentDefinitions) {
         this.schema = requireNonNull(schema);
         this.typeMappers = requireNonNull(typeMappers);
-        this.field = requireNonNull(field);
-
-        this.typeInfo = memoize(() -> maybeGetTypeInfo(parentObjectTypeDefinition, field));
-        this.objectTypeDefinition = memoize(() -> maybeFindObjectTypeDefinition(schema, typeInfo.get()));
-        this.path = memoize(() -> getTypeInfo().isList() ? emptyList() : appendToList(parentPath, getFieldAliasOrName(field)));
+        this.fragmentDefinitions = requireNonNull(fragmentDefinitions);
     }
 
-    static MappingContext of(TypeDefinitionRegistry schema, List<TypeMapper> typeMappers, ObjectTypeDefinition definition, Field field) {
-        return new MappingContext(schema, typeMappers, emptyList(), definition, field);
+    final Optional<TypeMapper> getTypeMapper() {
+        return maybeFindTypeMapper(typeMappers, getObjectTypeDefinition());
     }
 
-    MappingContext to(Field field) {
-        return new MappingContext(
-                this.schema,
-                this.typeMappers,
-                this.getPath(),
-                this.getObjectTypeDefinition(),
-                field);
+    private Optional<FragmentDefinition> maybeGetFragmentDefinition(String name) {
+        return fragmentDefinitions.stream().filter(fm -> fm.getName().equals(name)).findFirst();
     }
 
-    private List<String> getPath() {
-        return path.get();
+    final FragmentDefinition getFragmentDefinition(FragmentSpread fragmentSpread) {
+        return maybeGetFragmentDefinition(fragmentSpread.getName()).orElseThrow(IllegalStateException::new);
     }
 
-    String getSpringPath(String targetKey) {
-        return appendAsStream(getPath(), targetKey).map(p -> "['" + p + "']").collect(joining());
+    protected List<String> getPath() {
+        return Collections.emptyList();
     }
 
-    List<TypeMapper> getTypeMappers() {
-        return typeMappers;
+    final String getSpringPath(String targetKey) {
+        return concat(getPath().stream(), Stream.of(targetKey)).map(p -> "['" + p + "']").collect(joining());
     }
 
-    public TypeDefinitionRegistry getSchema() {
-        return schema;
+    boolean inList() {
+        return false;
     }
 
-    Field getField() {
-        return field;
+    protected abstract ObjectTypeDefinition getObjectTypeDefinition();
+
+    MappingContext forField(Field field) {
+        throw new IllegalStateException();
     }
 
-    TypeInfo getTypeInfo() {
-        return typeInfo.get().orElseThrow(IllegalStateException::new);
+    MappingContext forInlineFragment(InlineFragment inlineFragment) {
+        throw new IllegalStateException();
     }
 
-    ObjectTypeDefinition getObjectTypeDefinition() {
-        return objectTypeDefinition.get().orElseThrow(IllegalStateException::new);
+    static RootMappingContext rootContext(TypeDefinitionRegistry schema, List<TypeMapper> typeMappers) {
+        return new RootMappingContext(schema, typeMappers, emptyList());
     }
 
-    private static <T> Supplier<T> memoize(Supplier<T> supplier) {
-        return new MemoizingSupplier<>(supplier);
-    }
+    static final class RootMappingContext extends MappingContext {
 
-    private static <T> List<T> appendToList(List<T> list, T element) {
-        return appendAsStream(list, element).collect(toList());
-    }
+        RootMappingContext(MappingContext parentContext, List<FragmentDefinition> fragmentMappings) {
+            this(parentContext.schema, parentContext.typeMappers, fragmentMappings);
+        }
 
-    private static <T> Stream<T> appendAsStream(List<T> list, T element) {
-        return Stream.concat(list.stream(), Stream.of(element));
-    }
+        RootMappingContext(TypeDefinitionRegistry schema, List<TypeMapper> typeMappers, List<FragmentDefinition> fragmentMappings) {
+            super(schema, typeMappers, fragmentMappings);
+        }
 
+        FragmentDefinitionMappingContext forFragment(FragmentDefinition definition) {
+            return new FragmentDefinitionMappingContext(this, definition);
+        }
 
-    // copied from com.google.common.base.Suppliers.MemoizingSupplier
-    private static class MemoizingSupplier<T> implements Supplier<T>, Serializable {
-        final Supplier<T> delegate;
-        transient volatile boolean initialized;
-        // "value" does not need to be volatile; visibility piggy-backs
-        // on volatile read of "initialized".
-        transient T value;
+        OperationDefinitionMappingContext forOperationDefinition(OperationDefinition definition) {
+            return new OperationDefinitionMappingContext(this, definition);
+        }
 
-        MemoizingSupplier(Supplier<T> delegate) {
-            this.delegate = requireNonNull(delegate);
+        RootMappingContext withFragments(List<FragmentDefinition> fragmentMappings) {
+            return new RootMappingContext(this, fragmentMappings);
         }
 
         @Override
-        public T get() {
-            // A 2-field variant of Double Checked Locking.
-            if (!initialized) {
-                synchronized (this) {
-                    if (!initialized) {
-                        T t = delegate.get();
-                        value = t;
-                        initialized = true;
-                        return t;
-                    }
-                }
-            }
-            return value;
+        protected ObjectTypeDefinition getObjectTypeDefinition() {
+            throw new IllegalStateException();
+        }
+    }
+
+    static final class FragmentDefinitionMappingContext extends NodeMappingContext {
+
+        private final ObjectTypeDefinition objectTypeDefinition;
+
+        FragmentDefinitionMappingContext(MappingContext parentContext, FragmentDefinition definition) {
+            super(parentContext);
+            this.objectTypeDefinition = findFragmentObjectTypeDefinition(parentContext.schema, definition);
         }
 
         @Override
-        public String toString() {
-            return "Suppliers.memoize(" + delegate + ")";
+        public ObjectTypeDefinition getObjectTypeDefinition() {
+            return objectTypeDefinition;
         }
 
-        private static final long serialVersionUID = 0;
+        private static ObjectTypeDefinition findFragmentObjectTypeDefinition(TypeDefinitionRegistry schema, FragmentDefinition definition) {
+            return schema.getType(definition.getTypeCondition().getName()).map(ObjectTypeDefinition.class::cast).orElseThrow(IllegalStateException::new);
+        }
+    }
+
+    static abstract class NodeMappingContext extends MappingContext {
+
+        NodeMappingContext(MappingContext parentContext) {
+            super(parentContext);
+        }
+
+        @Override
+        NodeMappingContext forField(Field field) {
+            return new FieldMappingContext(this, field);
+        }
+
+        @Override
+        MappingContext forInlineFragment(InlineFragment inlineFragment) {
+            return new InlineFragmentMappingContext(this, inlineFragment);
+        }
+    }
+
+    static class OperationDefinitionMappingContext extends NodeMappingContext {
+        private final ObjectTypeDefinition objectTypeDefinition;
+
+        OperationDefinitionMappingContext(MappingContext parentContext, OperationDefinition operationDefinition) {
+            super(parentContext);
+            this.objectTypeDefinition = findOperationTypeDefinition(parentContext.schema, operationDefinition);
+        }
+
+        @Override
+        public ObjectTypeDefinition getObjectTypeDefinition() {
+            return objectTypeDefinition;
+        }
+
+        private static ObjectTypeDefinition findOperationTypeDefinition(TypeDefinitionRegistry schema, OperationDefinition op) {
+            return findOperationDefinitions(schema)
+                    .flatMap(maybeFindOperationTypeDefinition(op))
+                    .map(OperationTypeDefinition::getType)
+                    .flatMap(schema::getType)
+                    .map(BraidObjects::<ObjectTypeDefinition>cast)
+                    .orElseThrow(IllegalStateException::new);
+        }
+
+        private static Function<List<OperationTypeDefinition>, Optional<OperationTypeDefinition>> maybeFindOperationTypeDefinition(OperationDefinition op) {
+            return ops -> ops.stream().filter(isOperationTypeDefinitionForOperationType(op)).findFirst();
+        }
+
+        private static Predicate<OperationTypeDefinition> isOperationTypeDefinitionForOperationType(OperationDefinition op) {
+            return otd -> otd.getName().equalsIgnoreCase(op.getOperation().name());
+        }
+    }
+
+    private static class FieldMappingContext extends NodeMappingContext {
+
+        private final Field field;
+        private final List<String> parentPath;
+        private final TypeInfo typeInfo;
+        private final ObjectTypeDefinition objectTypeDefinition;
+
+        FieldMappingContext(MappingContext parentContext, Field field) {
+            super(parentContext);
+            this.field = requireNonNull(field);
+            this.parentPath = parentContext.getPath();
+            this.typeInfo = maybeGetTypeInfo(parentContext.getObjectTypeDefinition(), field).orElseThrow(IllegalStateException::new);
+            this.objectTypeDefinition = maybeFindObjectTypeDefinition(parentContext.schema, this.typeInfo).orElseThrow(IllegalStateException::new);
+        }
+
+        @Override
+        protected ObjectTypeDefinition getObjectTypeDefinition() {
+            return objectTypeDefinition;
+        }
+
+        @Override
+        protected List<String> getPath() {
+            return inList() ? emptyList() : concat(parentPath, getFieldAliasOrName(field));
+        }
+
+        @Override
+        boolean inList() {
+            return typeInfo.isList();
+        }
+    }
+
+    private static class InlineFragmentMappingContext extends NodeMappingContext {
+
+        private final List<String> parentPath;
+        private final ObjectTypeDefinition objectTypeDefinition;
+
+        InlineFragmentMappingContext(MappingContext parentContext, InlineFragment inlineFragment) {
+            super(parentContext);
+            this.parentPath = parentContext.getPath();
+            this.objectTypeDefinition = parentContext.schema.getType(inlineFragment.getTypeCondition().getName()).map(ObjectTypeDefinition.class::cast).orElseThrow(IllegalStateException::new);
+        }
+
+        @Override
+        protected ObjectTypeDefinition getObjectTypeDefinition() {
+            return objectTypeDefinition;
+        }
+
+        @Override
+        protected List<String> getPath() {
+            return parentPath;
+        }
+
+        @Override
+        boolean inList() {
+            return false;
+        }
     }
 }