+++ src/main/java/com/atlassian/braid/mapper/MapperOperation.java
@@ -8,7 +8,6 @@ import static com.atlassian.braid.mapper.MapperOperations.composed;
 /**
  * Implementation of a single {@link Mapper} operation, there is basically one implementation per mapper method.
  */
-@FunctionalInterface
 public interface MapperOperation extends BiConsumer<Map<String, Object>, Map<String, Object>> {
 
     void accept(Map<String, Object> input, Map<String, Object> output);
@@ -16,4 +15,6 @@ public interface MapperOperation extends BiConsumer<Map<String, Object>, Map<Str
     default MapperOperation andThen(MapperOperation after) {
         return composed(this, after);
     }
+
+//    MapperOperation toPath(String... path);
 }