+++ CHANGES.md
@@ -4,7 +4,7 @@ Changelog for graphql-braid
 (Unreleased)
 -------------------
 
-- 
+- Support GraphQL fragments when using document mapping
 
 0.9.1 (2018-04-09)
 -------------------